+++ src/model/location.rs
@@ -1,7 +1,7 @@
 use crate::earth::coordinate::Coordinate;
 
 pub trait Location: Clone {
-    fn get_elevation(&self) -> f64;
+    fn get_elevation(&self) -> i32;
     fn get_id(&self) -> String;
     fn get_lat(&self) -> f64;
     fn get_lat_as_string(&self) -> String;
